Snapshot
Seems OK-ish but looking at the colour histogram pairs it's obvious we
have "failed" because we didn't take that into account when chosing a
colour in pick_colour_from_colour_class. This isn't surprising, but this
seems a good point to try fixing that. (I had thought of this in the
original idea.)