Snapshot - it works, but not too well
I'm not sure if the output from this variant is better than the last;
perhaps it is a smidge better. However, I still think that to be honest
the output is not as good as from the original no-colour-class
implementation. That's not to say there's no value here, just that what
I currently have isn't very helpful - it may be some smallish tweaks
will make all the difference, e.g. see the TODOs.

Bedtime...