Snapshot - possibly better on PLAZA
With the current tweaked settings, that is. Not perfect.

I find myself wondering if for these purposes it would be better to
evaluate colour distance in HSV space and heavily weight the hue value.
Very quick poking on web suggests this isn't "right", but in this
particular application with such a limited palette and low resolution I
think it might be worth a try. I don't know if we'd do this only when
forming colour classes or also when picking best approximations during
the actual final stage pixel adjustments.