Try not to generate flat colours  - this is broken
We can sometimes allocate a perfect match colour to a near match and
then don't allow the perfect match to be used. This is just a snapshot
before I fiddle further. I'm trying to avoid flattening dithering into a
single three-pixel blur of the same closest match colour but this isn't
going to fly.