Don't be so harsh on colour reuse
Experimental - this is a bit better, I don't know if the magic number is
appropriately right at all. On PLAZA this is still probably worse than
the original no-colour-classification code (though it's not necessarily
the colour classification which is causing the degradation) but it does
have some small areas where it's a bit better.