Snapshot
This does produce output in z.png, not attempted to finish the code to
generate something which can be shown on BBC yet. The output is a bit
scrappy, it shows some improvement but also shows the danger of blindly
"doing the best" (a very poor shoddy best here due to poor algorithm as
it is) because where we have "spare" capacity in a line we will show
colours/detail which get missed on adjacent lines and it looks a bit
ropy.