Assemble code using beebasm
It's nice to be able to use : in comments. :-) This also shrinks the
size of showotf.bas so it works with PAGE=&1900 even with some of the
debug-type code left in. It would have been possible to split the BBC
BASIC assembly off into a loader program which then CHAINed "SSHOW", but
it's faster to pre-assemble and beebasm is a bit nicer than the BBC
BASIC assembler. If nothing else, it's nice to get any assembly errors
occuring at build time.